requirements.txt
loguru==0.5.3
sqlalchemy==1.3.22
gino==1.0.1
aiogram==2.11.0
aiohttp==3.7.2
python-dotenv
click==7.1.1
aiohttp_autoreload==0.0.1
